+++ photographer-master/src/main/java/com/fotos/photographer/master/service/repository/PhotoGMasterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhotoGMasterRepository extends JpaRepository<PhotographerMaster, Integer> {
+public interface PhotoGMasterRepository extends JpaRepository<PhotographerMaster, Long> {
 }
